pyproject.toml
[project]
name = "aa84-ai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "anthropic",
    "python-multipart",
    "pillow",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.115",
]